Mon, Apr 25, 2022, 11:15 PM +00:00

+++ project/app.py
@@ -244,6 +244,10 @@ def like():
         reply_id = request.form.get("reply_id")
         user_id = session["user_id"]
 
+        rows = db.execute("SELECT * FROM likes WHERE reply_id = ? AND user_id = ?", reply_id, user_id)
+        if len(rows) >= 1:
+            return apology("You've already liked this post!")
+
         db.execute("INSERT INTO likes (reply_id, user_id) VALUES (?, ?)", reply_id, user_id)
         db.execute("UPDATE replies SET likes = likes + 1 WHERE id = ?", reply_id)
 
@@ -260,6 +264,10 @@ def unlike():
     if request.method == "POST":
         reply_id = request.form.get("reply_id")
         user_id = session["user_id"]
+
+        rows = db.execute("SELECT * FROM likes WHERE reply_id = ? AND user_id = ?", reply_id, user_id)
+        if len(rows) >= 1:
+            return apology("You've already liked this post!")
 
         db.execute("DELETE FROM likes WHERE reply_id = ? AND user_id = ?", reply_id, user_id)
         db.execute("UPDATE replies SET likes = likes - 1 WHERE id = ?", reply_id)
